Add files via upload
Financial Performance Prediction task is highly creative and can be solved as a traditional ML problem with training models on features to predict targets. The dataset is contains a lot of static metadata characterizing the companies, along with dinamics of change of key financial data points across 11 consecutive quarters.

Alternatively contestants may use a time series approach based on target variables’ values recorded in the previous quarters.

One may also enrich the training/testing with macroeconomic statistics based on time frames present in the datasets - refer to data_dictionary.txt for mapping of quarters indicators in the dataset (Q0 - Q10) to the actual dates.